requirements.txt
mininet==2.3.0.dev6
networkx==3.4.2
numpy==2.1.3
ovs==3.4.1
pandas==2.2.3
pydantic==2.10.3
PyYAML==6.0.2
PyYAML==6.0.2
sortedcontainers==2.4.0
topohub==1.3
